+++ commands/src/main/java/net/apartium/cocoabeans/commands/parsers/CompoundParser.java
@@ -13,6 +13,7 @@ import net.apartium.cocoabeans.commands.requirements.*;
 import net.apartium.cocoabeans.structs.Entry;
 import org.jetbrains.annotations.ApiStatus;
 
+import java.lang.annotation.Annotation;
 import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Method;
@@ -27,6 +28,8 @@ public class CompoundParser<T> extends ArgumentParser<T> implements GenericNode 
     private final Map<Class<? extends ParserFactory>, ParserFactory> parserFactories = new HashMap<>();
     private final Map<Class<? extends ArgumentRequirementFactory>, ArgumentRequirementFactory> argumentRequirementFactories = new HashMap<>();
 
+    private final Map<Class<? extends Annotation>, RequirementFactory> externalRequirementFactories;
+
     private final CompoundParserBranchProcessor<T> compoundParserBranchProcessor;
 
     private final ArgumentMapper argumentMapper;
@@ -39,14 +42,33 @@ public class CompoundParser<T> extends ArgumentParser<T> implements GenericNode 
      * @param keyword keyword of the parser
      * @param clazz output class
      * @param priority priority
+     * @param argumentMapper argument mapper
+     * @param commandLexer command lexer
      */
     protected CompoundParser(String keyword, Class<T> clazz, int priority, ArgumentMapper argumentMapper, CommandLexer commandLexer) {
+        this(keyword, clazz, priority, argumentMapper, commandLexer, Collections.emptyMap());
+    }
+
+    /**
+     * Constructs a new parser
+     *
+     * @param keyword keyword of the parser
+     * @param clazz output class
+     * @param priority priority
+     * @param argumentMapper argument mapper
+     * @param commandLexer command lexer
+     * @param externalRequirementFactories external requirement factories
+     */
+    @ApiStatus.AvailableSince("0.0.38")
+    protected CompoundParser(String keyword, Class<T> clazz, int priority, ArgumentMapper argumentMapper, CommandLexer commandLexer, Map<Class<? extends Annotation>, RequirementFactory> externalRequirementFactories) {
         super(keyword, clazz, priority);
 
         this.argumentMapper = argumentMapper;
 
         this.commandLexer = commandLexer;
         this.compoundParserBranchProcessor = new CompoundParserBranchProcessor<>();
+
+        this.externalRequirementFactories = externalRequirementFactories;
 
         try {
             createBranch();
@@ -61,7 +83,7 @@ public class CompoundParser<T> extends ArgumentParser<T> implements GenericNode 
     }
 
     private void createBranch() throws IllegalAccessException {
-        RequirementSet requirementsResult = RequirementFactory.createRequirementSet(this, this.getClass().getAnnotations(), requirementFactories);
+        RequirementSet requirementsResult = RequirementFactory.createRequirementSet(this, this.getClass().getAnnotations(), requirementFactories, externalRequirementFactories);
         Map<String, ArgumentParser<?>> argumentParser = new HashMap<>();
 
         MethodHandles.Lookup lookup = MethodHandles.publicLookup();
@@ -77,7 +99,7 @@ public class CompoundParser<T> extends ArgumentParser<T> implements GenericNode 
                 continue;
 
             RequirementSet methodRequirements = new RequirementSet(
-                    RequirementFactory.createRequirementSet(this, method.getAnnotations(), requirementFactories),
+                    RequirementFactory.createRequirementSet(this, method.getAnnotations(), requirementFactories, externalRequirementFactories),
                     requirementsResult
             );